src/requirements.txt
grpcio>=1.45.0
grpc_interceptor>=0.14.1
grpcio-tools>=1.46.0
requests>=2.22.0
numpy>=1.19.5